sctcgbot/requirements.txt
aiohttp==3.12.15
discord.py==2.6.4
python-dotenv==1.1.1
requests==2.32.5